debundle: split realizability.rs into directory module
Pure file split of the ~3300-line realizability.rs (largest file in
the crate) into realizability/:

- esm_simulator.rs — EsmEvaluationSimulator + EsmIGraph +
  simulate_esm_post_order (the ECMA-262 Phase-2 evaluation-order
  simulator; the shared ordering itself stays in esm_import_order).
- incremental_quotient.rs — PartitionDelta/DeltaHandle journaling,
  QuotientOverlay + overlay machinery, IncrementalQuotient.
- tests.rs — the ~950-line in-module test module.
- mod.rs — check_realizability, RealizabilityIndex, SccDiagnosis/
  verdict types, and gate_perf_counters. gate_perf_counters stays:
  its pub(super) recording statics, use super::*, and the
  base_snapshot_stale shadow state on IncrementalQuotient are still
  entangled with index internals (re-confirmed; see CODE_REVIEW.md).

Cross-file internals are pub(super) — nothing new is visible outside
the realizability module. No behavior change.

BAZEL_TEST_INVOCATIONS=buildbuddy:c664b93a-399e-4de0-882d-dc19e7ab665d
