+++ devinfra/js/debundle/BUILD.bazel
@@ -307,7 +307,10 @@ rust_library(
         "chunk_factorization.rs",
         "esm_import_order.rs",
         "gate.rs",
-        "realizability.rs",
+        "realizability/esm_simulator.rs",
+        "realizability/incremental_quotient.rs",
+        "realizability/mod.rs",
+        "realizability/tests.rs",
         "report_builders.rs",
         "rollback_graph.rs",
         "validation.rs",
